tests/wpt/tests/navigation-api/navigation-history-entry/entries-after-server-redirect.html

<!doctype html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  // Do a redirected cross-document push navigation in the iframe.
  i.contentWindow.navigation.navigate("/common/redirect.py?location=/common/blank.html?redirected");
  await new Promise(resolve => i.onload = resolve);

  assert_equals(i.contentWindow.navigation.entries().length, 2);
  assert_equals(i.contentWindow.navigation.entries()[1], i.contentWindow.navigation.currentEntry);
  assert_equals(i.contentWindow.navigation.currentEntry.url, i.contentWindow.navigation.entries()[0].url + "?redirected");
}, "A navigation to a redirected page ends up on the page redirected to");
</script>
</body>
